+++ .gitignore
@@ -16,4 +16,4 @@ bin/
 /text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
 
-data/duke.txt
+data/chad.txt
