+++ README.md
@@ -1,4 +1,4 @@
-# Duke project template
+# chad project template
 
 This is a project template for a greenfield Java project. It's named after the Java mascot _Duke_. Given below are instructions on how to use it.
