+++ build.gradle
@@ -43,7 +43,7 @@ test {
 }
 
 application {
-    mainClass.set("duke.gui.Launcher")
+    mainClass.set("chad.gui.Launcher")
 }
 
 checkstyle {
@@ -51,7 +51,7 @@ checkstyle {
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "chad"
     archiveClassifier = null
     dependsOn("distZip", "distTar")
 }
