+++ src/main/java/chad/exceptions/ChadException.java
@@ -1,0 +1,10 @@
+package chad.exceptions;
+
+/**
+ * Exception unique to Chad.
+ */
+public class ChadException extends Exception {
+    public ChadException(String msg) {
+        super(msg);
+    }
+}
