+++ src/main/java/chad/gui/Launcher.java
@@ -1,0 +1,13 @@
+package chad.gui;
+
+import chad.Main;
+import javafx.application.Application;
+
+/**
+ * A Launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
