--- src/main/java/duke/command/ListCommand.java
@@ -1,26 +1,0 @@
-package duke.command;
-
-import duke.utility.Storage;
-import duke.utility.TaskList;
-import duke.utility.Ui;
-
-/**
- * Represents a command to show the current {@link TaskList}.
- */
-public class ListCommand extends Command {
-
-    /**
-     * Shows the current task list.
-     *
-     * @param list the task list
-     * @param ui Duke UI
-     * @param s Duke Storage
-     */
-    public void execute(TaskList list, Ui ui, Storage s) {
-        ui.listTasks(list.getList());
-    }
-
-    public boolean isExit() {
-        return false;
-    }
-}
