--- src/main/java/duke/exceptions/DukeException.java
@@ -1,10 +1,0 @@
-package duke.exceptions;
-
-/**
- * Exception unique to Duke.
- */
-public class DukeException extends Exception {
-    public DukeException(String msg) {
-        super(msg);
-    }
-}
