--- src/main/java/duke/gui/Launcher.java
@@ -1,13 +1,0 @@
-package duke.gui;
-
-import duke.Main;
-import javafx.application.Application;
-
-/**
- * A Launcher class to workaround classpath issues.
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
